+++ packages/lumx-react/src/components/menu-button/MenuButton.tsx
@@ -1,10 +1,12 @@
 import React, { type ForwardedRef } from 'react';
 
 import {
+    DEFAULT_PROPS,
     MenuButton as UI,
     type MenuButtonProps as UIProps,
     type MenuButtonVariantsProps,
     COMPONENT_NAME,
+    type MenuButtonVariant,
 } from '@lumx/core/js/components/Menu/MenuButton';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 import { ReactToJSX } from '@lumx/react/utils/type/ReactToJSX';
@@ -43,7 +45,7 @@ const TRIGGER_COMPONENTS = {
     'icon-button': IconButton,
     chip: Chip,
     link: Link,
-} as const;
+} satisfies Record<MenuButtonVariant, unknown>;
 
 /**
  * MenuButton component.
@@ -52,7 +54,7 @@ const TRIGGER_COMPONENTS = {
  * @return React element.
  */
 export const MenuButton = forwardRef<MenuButtonProps>((props, ref: ForwardedRef<HTMLElement>) => {
-    const { label, children, popoverProps, onOpen, variant = 'button', ...triggerProps } = props;
+    const { label, children, popoverProps, onOpen, variant = DEFAULT_PROPS.variant, ...triggerProps } = props;
 
     return UI(
         {
@@ -68,3 +70,4 @@ export const MenuButton = forwardRef<MenuButtonProps>((props, ref: ForwardedRef<
 });
 
 MenuButton.displayName = COMPONENT_NAME;
+MenuButton.defaultProps = DEFAULT_PROPS;