+++ packages/lumx-vue/src/components/menu-button/MenuButton.test.stories.tsx
@@ -1,13 +1,13 @@
 import { setup } from '@lumx/core/js/components/Menu/MenuButtonTestStories';
 
-import { MenuItem } from './MenuItem';
-import { MenuButton } from './MenuButton';
+import MenuItem from './MenuItem';
+import MenuButton from './MenuButton';
 
 const { meta, ...testStories } = setup({
     components: { MenuButton, MenuItem },
 });
 
-export default { ...meta, title: 'LumX components/menu/MenuButton/Tests' };
+export default { ...meta, title: 'LumX components/menu-button/MenuButton/Tests' };
 
 export const TabFromLastItemClosesAndMovesFocus = { ...testStories.TabFromLastItemClosesAndMovesFocus };
 export const ShiftTabFromFirstItemClosesAndReturnsToTrigger = {