+++ packages/lumx-vue/src/components/menu-button/index.ts
@@ -1,0 +1,3 @@
+export { default as MenuItem, type MenuItemProps, type MenuItemActionSlotProps } from './MenuItem';
+export { default as MenuButton, type MenuButtonProps } from './MenuButton';
+export { default as MenuDivider, type ListDividerProps as MenuDividerProps } from '../list/ListDivider';