+++ packages/lumx-vue/src/components/menu-button/useMenuOpen.ts
@@ -1,0 +1,13 @@
+import type { Ref } from 'vue';
+
+import { useMenuEvent } from './useMenuEvent';
+import { useMenuContext } from './context';
+
+export function useMenuOpen(): [Ref<boolean>, (open: boolean) => void] {
+    const { handle } = useMenuContext();
+    const isOpen = useMenuEvent('open', false);
+
+    const setIsOpen = (open: boolean) => handle.setOpen(open);
+
+    return [isOpen, setIsOpen];
+}